// src/pages/crm/main-page/types.ts
interface IFilterValues {
  query: string;
  is_criticism: {
    label: string;
    value: string;
  } | null;
  status: {
    label: string;
    value: string;
  } | null;
  target_ward: string | null;
  target_group: {
    label: string;
    value: string;
  } | null;
}

interface IFinalFeedbackModalForm {
  comment: string;
  file: File[] | [];
}

export type { IFilterValues, IFinalFeedbackModalForm };